hmt-shell/requirements.txt
pyyaml
requests
pymysql